+++ Adapters/AIAgents.Laboratory.Adapters.Agents/AIAgents.Laboratory.SemanticKernel.Adapters.csproj
